requirements.txt
Pyrogram
requests
aiohttp
Tgcrypto
asyncio
bs4
telethon
Pillow
motor
dnspython
pymongo
umongo
psutil
aiofiles
DateTime
NumPy
glitch_this
opencv-python-headless
